% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bounds.R
\name{bucklandbounds.cells}
\alias{bucklandbounds.cells}
\title{bucklandbounds.cells}
\usage{
bucklandbounds.cells(thetaa, var_est, w, sparsemat, cellsix, critval)
}
\arguments{
\item{thetaa}{Stacked relative risk estimates for each cell.}

\item{var_est}{Variances for each single cluster model.}

\item{w}{Likelihood-weights for potential clusters.}

\item{sparsemat}{Large sparse matrix where columns are potential clusters and rows are space-time locations.}

\item{cellsix}{Indices of the cells to calculate bounds for.}

\item{critval}{Critical value associated with the conf.level.}
}
\value{
List: lower bound estimate, stacked cluster relative risk estimate, upper bound estimate for each of the cells in \code{cellsix}
}
\description{
Calculate confidence bounds for stacked cluster relative risk estimates based on Buckland et al. for specific cells (given by \code{cellsix})
}
