% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustack.R
\name{bylocation}
\alias{bylocation}
\title{bylocation}
\usage{
bylocation(Lik, Lambda_dense, sparsemat, maxclust)
}
\arguments{
\item{Lik}{Likelihood for each potential cluster.}

\item{Lambda_dense}{Initial estimates for relative risk for each potential cluster;\eqn{\Lambda}.}

\item{sparsemat}{Large sparse matrix where rows are potential clusters and columns are space-time locations.}

\item{maxclust}{Maximum number of clusters to be detected.}
}
\value{
List. First element is a Weighted relative risks for identified locations; second element is a large matrix of weights for each iteration; third element are the maximum location IDs.
}
\description{
Cluster detection based on maximum location.
}
