% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bounds.R
\name{calcbounds.cells}
\alias{calcbounds.cells}
\title{calcbounds.cells}
\usage{
calcbounds.cells(
  id_ic,
  IC,
  res,
  byloc = FALSE,
  Ex,
  Yx,
  w,
  thetaa,
  thetai,
  sparsemat,
  cellsix,
  conf.level = 0.95,
  overdisp.est
)
}
\arguments{
\item{id_ic}{The number of clusters identified either by BIC (QBIC) or AIC (QAIC).}

\item{IC}{Information criterion used. Currently available for BIC (QBIC) and AIC (QAIC). This should match \code{id_ic} argument.}

\item{res}{Resultant object from \code{detectclusters()}.}

\item{byloc}{Boolean. Use \code{TRUE} when stacking by location. Use \code{FALSE} when stacking by potential cluster.}

\item{Ex}{Expected counts (unstandardized) for each cell.}

\item{Yx}{Observed counts for each cell.}

\item{w}{Likelihood-weights for all potential clusters.}

\item{thetaa}{Stacked relative risk estimate for the cluster(s).}

\item{thetai}{Individual relative risk estimates for each potential cluster.}

\item{sparsemat}{Large sparse matrix where rows are potential clusters and columns are space-time locations.}

\item{cellsix}{Indices of the cells to calculate bounds for.}

\item{conf.level}{Confidence level for the interval. Default is 0.95.}

\item{overdisp.est}{Overdispersion estimate.}
}
\value{
Returns large list of confidence bounds and stacked estimates in addition to timings for each of the confidence bounds methods.
}
\description{
Helper function for \code{calcbounds()}.
}
