% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bounds.R
\name{calcbounds}
\alias{calcbounds}
\title{calcbounds}
\usage{
calcbounds(
  id_ic,
  IC,
  res,
  byloc,
  Ex,
  Yx,
  cellsix = NULL,
  sparsemat,
  conf.level = 0.95,
  overdisp.est = NULL
)
}
\arguments{
\item{id_ic}{The number of clusters identified either by BIC (QBIC) or AIC (QAIC).}

\item{IC}{Information criterion used. Currently available for BIC (QBIC) and AIC (QAIC). This should match \code{id_ic} argument.}

\item{res}{Resultant object from \code{detectclusters()}.}

\item{byloc}{Boolean. Use \code{TRUE} when stacking by location. Use \code{FALSE} when stacking by potential cluster.}

\item{Ex}{Expected counts (unstandardized) for each cell.}

\item{Yx}{Observed counts for each cell.}

\item{cellsix}{Indices of the cells to calculate bounds for. If you you want to calculate bounds for all cells, then enter these as a vector (ex: 1:208).}

\item{sparsemat}{Large sparsematrix where rows are potential clusters and columns are space-time locations.}

\item{conf.level}{Confidence level for the interval. Default is 0.95.}

\item{overdisp.est}{Overdispersion estimate.}
}
\description{
Calculates lower and upper confidence bounds for stacked relative risk estimates using Buckland, Burnham and Anderson, and MATA methods.
}
\details{
The confidence bounds methods returned are:
\itemize{
\item \code{buckTlog}: Stacked confidence bounds based on Buckland et al.; natural log-transformed.
\item \code{baTlog}: Model-average weighted confidence bounds based on Burnham and Anderson; natural log-transformed.
\item \code{mataTlog}: Model-average tail area (MATA) intervals based on Turek et al.; natural log-transformed.

}
}
