% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustack.R
\name{clusterselect}
\alias{clusterselect}
\title{clusterselect}
\usage{
clusterselect(
  wLambda,
  Yx,
  Ex,
  model,
  maxclust,
  numCenters,
  Time,
  isoverdisp,
  overdisp.est
)
}
\arguments{
\item{wLambda}{Stacked estimates for relative risks in all locations.}

\item{Yx}{Vector of observed counts.}

\item{Ex}{Vector of expected counts.}

\item{model}{A string specifying which model to use, Poisson or binomial. For Poisson, specify \code{"poisson"} and both the Poisson and quasi-Poisson model results are returned. For binomial, specify \code{"binomial"}.}

\item{maxclust}{Maximum number of clusters to consider.}

\item{numCenters}{Number of polygon centroids (total number of locations).}

\item{Time}{Number of time periods.}

\item{isoverdisp}{Boolean. \code{TRUE} indicates a quasi-Poisson or quasi-binomial model that accounts for overdispersion. \code{FALSE} indicates a Poisson or binomial model without adjustment for overdispersion or underdispersion.}

\item{overdisp.est}{Overdispersion (or underdispersion) parameter estimate.}
}
\value{
Returns a list with the loglikelihood, selection by (Q)BIC, selection by (Q)AIC, and selection by (Q)AICc.
}
\description{
Select the optimal number of (overlapping) clusters using either information criteria.
}
