% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bounds.R
\name{f_LB}
\alias{f_LB}
\title{f_LB}
\usage{
f_LB(theta, sesi, thetaii, w, alpha2)
}
\arguments{
\item{theta}{The parameter to be optimized.}

\item{sesi}{Standard error for each cell.}

\item{thetaii}{Model estimated relative risk for each cell.}

\item{w}{Likelihood-weights for potential clusters.}

\item{alpha2}{Alpha value/2}
}
\description{
Function over which the lower bound root is sought for MATA bounds
}
