% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jbc2.R
\docType{data}
\name{jbc2}
\alias{jbc2}
\title{Japanese Breast Cancer Data 2 (JBC2)}
\format{
An object of class "data.frame" with 1040 rows and 9 variables:
\itemize{
\item Observed: count of the number of observed incident breast cancer deaths in each centroid-time period.
\item Expected: expected number of incident breast cancer deaths in each centroid-time period (age-standardized).
\item Time: time period at which each centroid location was measured (there are 5 total time periods, factor).
}
}
\usage{
jbc2
}
\description{
Data on breast cancer incidence in prefects in Japan across 5 time periods. A fake cluster has been placed in center 150, with a radius of < 15km, in periods 3 through 5. A relative risk of 2 was used to create this cluster.
}
\examples{
data(jbc2)
}
\keyword{datasets}
