% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustack.R
\name{likweights}
\alias{likweights}
\title{likeweights}
\usage{
likweights(liki)
}
\arguments{
\item{liki}{Likelihood for each potential cluster.}
}
\value{
A vector of weights.
}
\description{
Generate likelihood-based initial weights.
}
