% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bounds.R
\name{mata_ZUB}
\alias{mata_ZUB}
\title{mata_ZUB}
\usage{
mata_ZUB(theta, sesi, thetaii, w)
}
\arguments{
\item{theta}{The parameter to be optimized.}

\item{sesi}{Standard error for each cell.}

\item{thetaii}{Model estimated relative risk for each cell.}

\item{w}{Likelihood-weights for potential clusters.}
}
\description{
Calculates upper bound using MATA confidence bounds for a given theta
}
