% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bounds.R
\name{matabounds_log.cells}
\alias{matabounds_log.cells}
\title{matabounds_log.cells}
\usage{
matabounds_log.cells(var_est, thetai, thetaa, w, cellsix, alpha2)
}
\arguments{
\item{var_est}{Variances for each single cluster model.}

\item{thetai}{Large matrix of model estimates and background relative risks (set to 1) where rows are correspond to each single cluster model and columns correspond to each cell in space (and/or time).}

\item{thetaa}{Stacked relative risk estimate for the cluster(s).}

\item{w}{Likelihood-weights for potential clusters.}

\item{cellsix}{Indices of the cells to calculate bounds for.}

\item{alpha2}{Alpha value/2}
}
\description{
Calculate confidence bounds for stacked cluster relative risk estimates using MATA (model-averaged tail area) bounds for specific cells (given by \code{cellsix})
}
