% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustack.R
\name{overdisp}
\alias{overdisp}
\title{overdisp}
\usage{
overdisp(offset_reg, overdispfloor = TRUE)
}
\arguments{
\item{offset_reg}{Object of class glm or lm.}

\item{overdispfloor}{Boolean. When \code{TRUE}, overdispersion parameter (\eqn{\phi}) is limited to not be less than 1. If \code{FALSE}, underdispersion can be estimated.}
}
\value{
An estimate of \eqn{\phi} (overdispersion or underdispersion)
}
\description{
Calculate the overdispersion parameter (\eqn{\phi}),
}
