% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustack.R
\name{poisLik}
\alias{poisLik}
\title{poisLik}
\usage{
poisLik(Ex, Yx, sparsemat)
}
\arguments{
\item{Ex}{Vector of expected counts.}

\item{Yx}{Vector of observed counts.}

\item{sparsemat}{Large sparse matrix where rows are potential clusters and columns are space-time locations.}
}
\value{
Likelihood for each potential cluster.
}
\description{
Poisson-based LRT statistic comparing cluster model to null model.
}
