% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bounds.R
\name{selectuniq}
\alias{selectuniq}
\title{selectuniq}
\usage{
selectuniq(uniqRRs, background)
}
\arguments{
\item{uniqRRs}{Matrix from which to select unique elements.}

\item{background}{Background from which elements will be different (typically 1 or zero)}
}
\description{
Selects unique elements from a matrix that are different from the background by row
}
