% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jbc2.R, R/utmJapan.R
\docType{data}
\name{utmJapan}
\alias{utmJapan}
\title{UTM coordinates}
\format{
An object of class "data.frame" with 208 rows and 3 variables:
\itemize{
\item id: identifier for each centroid location.
\item utmx: easting (x-coordinate).
\item utmy: northing (y-coordinate).
}

An object of class "data.frame" with 208 rows and 3 variables:
\itemize{
\item id: identifier for each centroid location.
\item utmx: easting (x-coordinate).
\item utmy: northing (y-coordinate).
}
}
\usage{
utmJapan

utmJapan
}
\description{
UTM coordinates for Japanese Breast Cancer dataset (JBC)

UTM coordinates for Japanese Breast Cancer dataset (JBC)
}
\examples{
data(utmJapan)
data(utmJapan)
}
\keyword{datasets}
